Refactor custom app to share state across pages
