+++ client/pages/view/[customSetId].tsx
@@ -5,66 +5,31 @@ import { jsx } from '@emotion/core';
 import { NextPage } from 'next';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import { useQuery } from '@apollo/react-hooks';
 
 import MobileSetBuilder from 'components/mobile/SetBuilder';
 import { mediaStyles, Media } from 'components/common/Media';
-import CustomSetQuery from 'graphql/queries/customSet.graphql';
-import {
-  customSet as CustomSetQueryType,
-  customSetVariables,
-} from 'graphql/queries/__generated__/customSet';
+
 import { CustomSetHead } from 'common/wrappers';
 import ClassicSetBuilder from 'components/desktop/ClassicSetBuilder';
 import {
   ClassicContext,
   useClassic,
   EditableContext,
-  appliedBuffsReducer,
   CustomSetContext,
-  getStatsFromCustomSet,
-  getStatsFromAppliedBuffs,
-  combineStatsWithBuffs,
 } from 'common/utils';
 import DesktopLayout from 'components/desktop/Layout';
 import MobileLayout from 'components/mobile/Layout';
-import { AppliedBuffActionType } from 'common/types';
 import ErrorPage from '../_error';
 
 const Index: NextPage = () => {
   const router = useRouter();
   const { customSetId } = router.query;
 
-  const { data: customSetData, loading } = useQuery<
-    CustomSetQueryType,
-    customSetVariables
-  >(CustomSetQuery, { variables: { id: customSetId }, skip: !customSetId });
-
-  const customSet = customSetData?.customSetById || null;
+  const { customSet } = React.useContext(CustomSetContext);
 
   const [isClassic, setIsClassic] = useClassic();
-  const [appliedBuffs, dispatch] = React.useReducer(appliedBuffsReducer, []);
 
-  const statsFromCustomSet = React.useMemo(
-    () => getStatsFromCustomSet(customSet),
-    [customSet],
-  );
-
-  const statsFromAppliedBuffs = React.useMemo(
-    () => getStatsFromAppliedBuffs(appliedBuffs),
-    [appliedBuffs],
-  );
-
-  const statsFromCustomSetWithBuffs = React.useMemo(
-    () => combineStatsWithBuffs(statsFromCustomSet, statsFromAppliedBuffs),
-    [statsFromCustomSet, statsFromAppliedBuffs],
-  );
-
-  React.useEffect(() => {
-    dispatch({ type: AppliedBuffActionType.CLEAR_ALL });
-  }, [customSetId]);
-
-  if (!customSet && !loading) {
+  if (customSetId && !customSet) {
     return <ErrorPage statusCode={404} />;
   }
 
@@ -79,29 +44,18 @@ const Index: NextPage = () => {
           />
         </Head>
         <CustomSetHead customSet={customSet} />
-        <CustomSetContext.Provider
-          value={{
-            dispatch,
-            appliedBuffs,
-            customSet,
-            statsFromCustomSet,
-            statsFromAppliedBuffs,
-            statsFromCustomSetWithBuffs,
-          }}
-        >
-          <EditableContext.Provider value={false}>
-            <Media lessThan="xs">
-              <MobileLayout>
-                <MobileSetBuilder customSet={customSet} />
-              </MobileLayout>
-            </Media>
-            <Media greaterThanOrEqual="xs" css={{ height: '100%' }}>
-              <DesktopLayout showSwitch={false}>
-                <ClassicSetBuilder customSet={customSet} />
-              </DesktopLayout>
-            </Media>
-          </EditableContext.Provider>
-        </CustomSetContext.Provider>
+        <EditableContext.Provider value={false}>
+          <Media lessThan="xs">
+            <MobileLayout>
+              <MobileSetBuilder customSet={customSet} />
+            </MobileLayout>
+          </Media>
+          <Media greaterThanOrEqual="xs" css={{ height: '100%' }}>
+            <DesktopLayout showSwitch={false}>
+              <ClassicSetBuilder customSet={customSet} />
+            </DesktopLayout>
+          </Media>
+        </EditableContext.Provider>
       </div>
     </ClassicContext.Provider>
   );
